Moved cache to utils and renamed endpoints
